integrates\refac(back): #6065.33  fix mypy errors
- fix mypy api groups  module

+++ integrates/back/src/api/resolvers/group/authors.py
@@ -58,7 +58,7 @@ async def resolve(
     )
     if group_tier.upper() == GroupTier.SQUAD.value:
         prices: Dict[str, Price] = await billing_domain.get_prices()
-        current_spend = total * prices["squad"].amount / 100
+        current_spend = int(total * prices["squad"].amount / 100)
     else:
         current_spend = 0
